Move `clean` related functionality and tests into `clean.go`/`clean_test.go`
